[#12 Use float instead of double]

+++ src/BRDFs/GlossySpecular.cpp
@@ -108,11 +108,11 @@ GlossySpecular::set_samples(const int num_samples, const float exp) {
 RGBColor
 GlossySpecular::f(const ShadeRec &sr, const Vector3D &wo, const Vector3D &wi) const {
     RGBColor    L;
-    double      ndotwi = sr.normal * wi;
+    float      ndotwi = sr.normal * wi;
     Vector3D    r(-wi + 2.0 * sr.normal * ndotwi); // mirror reflection direction
-    double      rdotwo = r * wo;
+    float      rdotwo = r * wo;
 
-    if (rdotwo > 0.0) {
+    if (rdotwo > 0.0f) {
         L = ks * cs * pow(rdotwo, exp);
     }
 
@@ -125,18 +125,18 @@ GlossySpecular::f(const ShadeRec &sr, const Vector3D &wo, const Vector3D &wi) co
 RGBColor
 GlossySpecular::sample_f(const ShadeRec& sr, const Vector3D& wo, Vector3D& wi, float& pdf) const {
 
-    double ndotwo = sr.normal * wo;
+    float ndotwo = sr.normal * wo;
     Vector3D r = -wo + 2.0 * sr.normal * ndotwo;     // direction of mirror reflection
 
     Vector3D w = r;
-    Vector3D u = Vector3D(0.00424, 1, 0.00764) ^ w;
+    Vector3D u = Vector3D(0.00424f, 1.0f, 0.00764f) ^ w;
     u.normalize();
     Vector3D v = u ^ w;
 
     Point3D sp = sampler_ptr->sample_hemisphere();
     wi = sp.x * u + sp.y * v + sp.z * w;            // reflected ray direction
 
-    if (sr.normal * wi < 0.0) {                     // reflected ray is below tangent plane
+    if (sr.normal * wi < 0.0f) {                     // reflected ray is below tangent plane
         wi = -sp.x * u - sp.y * v + sp.z * w;
     }
