+++ src/GeometricObjects/ConvexPartCylinder.h
@@ -22,7 +22,7 @@ public:
 
     ConvexPartCylinder() = default;
 
-    ConvexPartCylinder(const double bottom, const double top, const double r, double phi1, double phi2);
+    ConvexPartCylinder(const float bottom, const float top, const float r, float phi1, float phi2);
 
     ~ConvexPartCylinder() = default;
 
@@ -40,19 +40,19 @@ public:
     clone() const override;
 
     bool
-    hit(const Ray& ray, double& tmin, ShadeRec& sr) const override;
+    hit(const Ray& ray, float& tmin, ShadeRec& sr) const override;
 
     BBox
     get_bounding_box() const override;
 
 private:
 
-    double y0         = -1.0;
-    double y1         = 1.0;
-    double radius     = 1.0;
-    double inv_radius = 1.0;
-    double phi_min    = 0.0;
-    double phi_max    = 180.0;
+    float y0         = -1.0f;
+    float y1         = 1.0f;
+    float radius     = 1.0f;
+    float inv_radius = 1.0f;
+    float phi_min    = 0.0f;
+    float phi_max    = 180.0f;
 };
 
 #endif
